kurt

+++ datasets/music/_2npy.py
@@ -34,8 +34,8 @@ np_arr = np.array(arr)
 
 # 88/6/6 split 
 length = len(np_arr)
-train_size = np.floor(length * .88) # train
-test_size = np.floor(length * .06) # test
+train_size = int(np.floor(length * .88)) # train
+test_size = int(np.floor(length * .06)) # test
 
 np.save('all_kurt.npy', np_arr)
 np.save('kurt_train.npy', np_arr[:train_size])
